+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/standardtime/primitivevalue/LimitThreeMonths.java
@@ -11,7 +11,7 @@ import nts.arc.primitive.constraint.TimeRange;
 @TimeRange(max="2232:00", min = "00:00")
 public class LimitThreeMonths extends TimeDurationPrimitiveValue<LimitThreeMonths> {
 
-	public LimitThreeMonths(Long rawValue) {
+	public LimitThreeMonths(int rawValue) {
 		super(rawValue);
 	}
