+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/common/time/AttendanceTime.java
@@ -18,7 +18,7 @@ public class AttendanceTime extends TimeDurationPrimitiveValue<AttendanceTime> {
 	 * @param rawValue
 	 *            the raw value
 	 */
-	public AttendanceTime(Long rawValue) {
+	public AttendanceTime(Integer rawValue) {
 		super(rawValue);
 	}
 }
